depgraph replaces deptree
depgraph sorts modules in topological order (aka resolves dependencies).
This "fixes" sorting issues that existed in the previous 'deptree'
implementation (which also created a graph).

This commit also adds the --depgraph/--revdep actions, which visualize the
edges of the depgraph as follows:

* Module_A ==> Module_B

   Module_A requires Module_B (--depgraph output)

* Module_A <== Module_B

   Module_A is required by Module_B (--revdep output)

The --deptree action has been superseded by the --deplist action, which prints
the order in which modules would be written/processed.
The module order is not stable (can vary from run to run) due to the use of
unordered data collection types (set()). It should always be correct, though.

This commit raises shlibcc's version to 0.0.5.