shlibcclib/main: allow direct imports in --depfile
The file passed via --depfile can now contain relative or absolute paths to
other modules or scripts, which are handled as dependencies:

* ./<path>,
  ../<path> -- path is relative to depfile's dir
* /<path>   -- path is relative to --shlib-dir's parent dir
* //<path>  -- path is absolute

In any case, path must not be a directory.

Note that this feature is intentionally restricted to --depfile.
'direct imports' should not be used unless you want to extend an existent
script without modifying it.
Even then, it's (more or less) only useful for creating standalone scripts.