deptable, main: recognize blocker files
Blocker files is a feature that prevents shlibcc from including directories
that contain a file/dir named "block_CC". However, it does not prevent
shlibcc from including modules from such directories (which is OK).
Several actions exist to handle blocker files - ignore, exclude (quiet/verbose),
exit with error, continue with warning. The default is 'exit with error' as it's
generally not a good idea to include dirs with blockers.